requirements.txt
PyYAML==6.0.2
boto3==1.35.88
botocore==1.35.88
cmlutils
fontTools==4.55.3
pyats[full]
sshuttle
